pos_idx and neg_idx to default float tensor

+++ layers/modules/multibox_loss.py
@@ -143,8 +143,8 @@ class MultiBoxLoss(nn.Module):
         neg_idx = neg.unsqueeze(2).expand_as(conf_data)
         
         if self.use_gpu == 2:
-            pos_idx = pos_idx.to(device, dtype=torch.long)
-            neg_idx = neg_idx.to(device, dtype=torch.long)
+            pos_idx = pos_idx.to(device)
+            neg_idx = neg_idx.to(device)
             
         conf_p = conf_data[(pos_idx+neg_idx).gt(0)].view(-1, self.num_classes)
         targets_weighted = conf_t[(pos+neg).gt(0)]
